add kill process confirm

+++ killOneProcess.sh
@@ -12,7 +12,7 @@ function killProcess(){
         return
     else
         pid=$(ps -ef | grep -E $processName | grep -v "grep" | awk '{print $2}')
-		pidArray=$(ps -ef | grep -E $processName | grep -v "grep" | awk '{print $0}')
+        pidArray=$(ps -ef | grep -E $processName | grep -v "grep" | awk '{print $0}')
         pidCount=$(ps -ef | grep -E $processName | grep -v "grep" | awk '{print $2}' | wc -l)
        # printf "Process pid is: "$pid"\n"
     fi
@@ -25,9 +25,15 @@ function killProcess(){
     else
 	if [[ ${pidCount} == 1 ]]
 	then
-            printf "Process pid is: "$pid"\n"
-            kill -9  $pid
-	    exit $?
+	    printf "$pidArray\n"
+	    read -p "Please confirm the pid(y/n):" res
+	    if [ $res = "y" ]
+	    then
+                kill -9  $pid
+	        exit $?
+	    else
+		return
+	    fi
 	else
 	    printf "Find these process:\n"
 	    printf "${pidArray[@]}\n"
